backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.1.13
langchain-google-genai>=1.0.1
python-dotenv>=1.0.1
pydantic>=2.6.4
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9